Fix lint and unit tests (#658)

+++ crates/gem_chain_rpc/src/xrp.rs
@@ -2,8 +2,8 @@ use async_trait::async_trait;
 use std::error::Error;
 
 use crate::{ChainAssetsProvider, ChainBlockProvider, ChainStakeProvider, ChainTokenDataProvider, ChainTransactionsProvider};
-use gem_xrp::rpc::{XRPClient, XRPMapper};
 use gem_client::Client;
+use gem_xrp::rpc::{XRPClient, XRPMapper};
 use primitives::{Asset, AssetBalance, Chain, Transaction};
 
 pub struct XRPProvider<C: Client> {
@@ -23,7 +23,7 @@ impl<C: Client> ChainBlockProvider for XRPProvider<C> {
     }
 
     async fn get_latest_block(&self) -> Result<i64, Box<dyn Error + Send + Sync>> {
-        Ok(self.client.get_ledger_current().await?.ledger_current_index as i64)
+        Ok(self.client.get_ledger_current().await?.ledger_current_index)
     }
 
     async fn get_transactions(&self, block_number: i64) -> Result<Vec<Transaction>, Box<dyn Error + Send + Sync>> {
@@ -55,4 +55,4 @@ impl<C: Client> ChainTransactionsProvider for XRPProvider<C> {
 }
 
 #[async_trait]
-impl<C: Client> ChainStakeProvider for XRPProvider<C> { }
+impl<C: Client> ChainStakeProvider for XRPProvider<C> {}
